applications/text_to_speech/pyproject.toml
[tool.poetry]
name = "speech-recognition"
version = "0.1.0"
description = ""
authors = ["Jacek Lipiec <jacek.lipiec.bc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
openai-whisper = "20240930"
torch = { version = "2.4.1", source = "pytorch-gpu" }
torchaudio = { version = "2.4.1", source = "pytorch-gpu" }
numpy = "1.22.0"
Flask = "^3.0.3"
ffmpeg-python = "^0.2.0"
python-config = { path = "../../libraries/python_config", develop = true }
melotts = { git = "https://github.com/myshell-ai/MeloTTS.git", rev = "5b538481e24e0d578955be32a95d88fcbde26dc8" }
myshell-openvoice = { git = "https://github.com/myshell-ai/OpenVoice.git", rev = "f3cf835540572ade1460c8952f39d53e4f7952df" }

[[tool.poetry.packages]]
include = "speech_recognition"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
